frontend: Remove deprecated Editors
Removed usages of deprecated EntityModelCheckBoxOnlyEditor,
EntityModelPasswordBoxEditor, EntityModelTextAreaLabelEditor and
repaced with generic equivalents.

Also generified related Models.
- ConfirmationModel
- LoginModel
- UserPortalLoginModel
- EventNotificationModel

Storage:
- ClusterGlusterHookListModel
- GlusterHookContentModel
- GlusterHookResolveConflictsModel
- BrickAdvancedDetailsModel
- BrickPropertiesModel
- StorageModel
- MemoryStatisticsModel
- NfsStorageModel

Infra:
- HostModel
- InstallModel
- MultipleHostsModel

Virt:
- AbstractDiskModel
- DiskModel
- NewDiskModel
- SnapshotModel

Network:
- ProviderModel
- HostNetworkProviderModel
- NeutronAgentModel

+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/models/vms/EditDiskModel.java
@@ -64,7 +64,7 @@ public class EditDiskModel extends AbstractDiskModel
             }, getHash()), storageDomainId);
         } else {
             LunDisk lunDisk = (LunDisk) getDisk();
-            getSize().setEntity(lunDisk.getLun().getDeviceSize());
+            getSize().setEntity((long) lunDisk.getLun().getDeviceSize());
             getIsInternal().setEntity(false);
             getSizeExtend().setIsAvailable(false);
         }
